--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Final/Final/TeleOp/teleop_interface.java
@@ -1,26 +1,0 @@
-package org.firstinspires.ftc.teamcode.Final.Final.TeleOp;
-
-
-import org.firstinspires.ftc.teamcode.mainEnum;
-
-public interface teleop_interface {
-    void initializeArms();
-    void initializeWheels();
-
-    void setDirectionArms();
-    void setDirectionWheels();
-
-    void setBrakesArms();
-    void setBrakesWheels();
-
-    void telemetry();
-
-    void movement(double vertical, double strafe, double turn);
-    void finalMovement();
-
-    void arm(mainEnum state, double speed);
-    void finalArm();
-
-    void claw(mainEnum state, double speed, int pos);
-    void finalGrabber();
-}
