+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware.java
@@ -1,12 +1,11 @@
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.qualcomm.robotcore.hardware.Servo;
 
-import org.firstinspires.ftc.teamcode.oldCrap.newAuto.main.newAuto_interface;
+import org.firstinspires.ftc.teamcode.Old.newAuto.main.newAuto_interface;
 
 public class hardware {
     // Arm components
