--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/oldCrap/newAuto/teaching/teach/complexAuto/baseAutoEnum.java
@@ -1,10 +1,0 @@
-package org.firstinspires.ftc.teamcode.oldCrap.newAuto.teaching.teach.complexAuto;
-
-public enum baseAutoEnum {
-    FORWARD,
-    BACKWARD,
-    STRAFE_LEFT,
-    STRAFE_RIGHT,
-    TURN_RIGHT,
-    TURN_LEFT
-}
