--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/oldCrap/newAuto/teaching/teach/complexAuto/baseHardware.java
@@ -1,14 +1,0 @@
-package org.firstinspires.ftc.teamcode.oldCrap.newAuto.teaching.teach.complexAuto;
-
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.DcMotor;
-
-public class baseHardware {
-    public DcMotor frontLeft;
-    public DcMotor frontRight;
-    public DcMotor backLeft;
-    public DcMotor backRight;
-    public DcMotor lift;
-
-    public CRServo claw;
-}
